Update auto and clean up unnecessary files
