+++ frontend/src/app/supportAdmin/DeviceType/DeviceForm.tsx
@@ -135,15 +135,17 @@ const DeviceForm = (props: Props) => {
                   />
                 </div>
               </div>
-              <div className="tablet:grid-col">
-                <TextInput
-                  label="Model"
-                  name="model"
-                  value={device?.model}
-                  onChange={onChange}
-                  disabled={!device}
-                  required
-                />
+              <div className="grid-row grid-gap">
+                <div className="tablet:grid-col">
+                  <TextInput
+                    label="Model"
+                    name="model"
+                    value={device?.model}
+                    onChange={onChange}
+                    disabled={!device}
+                    required
+                  />
+                </div>
               </div>
               <div className="grid-row grid-gap">
                 <div className="tablet:grid-col">
